move `tablev2`-> `table/v2`
