+++ clients/admin-ui/src/features/common/table/v2/RowSelectionBar.tsx
